Tools: Move raw message reader into msg module.
