+++ tools/bl.c
@@ -22,12 +22,10 @@
 #include <string.h>
 #include <errno.h>
 #include <stdio.h>
-#include <time.h>
 
 #include <termios.h>
 #include <unistd.h>
 #include <fcntl.h>
-#include <poll.h>
 
 #include "../src/error.h"
 #include "../src/util.h"
@@ -77,127 +75,11 @@ static inline bool read_sized_uint(
 	return true;
 }
 
-
-
-static inline int64_t time_diff_ms(
-		struct timespec *time_start,
-		struct timespec *time_end)
-{
-	return ((time_end->tv_sec - time_start->tv_sec) * 1000 +
-		(time_end->tv_nsec - time_start->tv_nsec) / 1000000);
-}
-
-ssize_t bl_read(int fd, void *data, size_t size, int timeout_ms)
-{
-	struct timespec time_start;
-	struct timespec time_end;
-	size_t total_read = 0;
-	int ret;
-
-	ret = clock_gettime(CLOCK_MONOTONIC, &time_start);
-	if (ret == -1) {
-		fprintf(stderr, "Failed to read time: %s\n",
-				strerror(errno));
-		return -errno;
-	}
-
-	while (total_read != size && !bl_sig_killed) {
-		ssize_t chunk_read;
-		struct pollfd pfd = {
-			.fd = fd,
-			.events = POLLIN,
-		};
-
-		ret = poll(&pfd, 1, timeout_ms);
-		if (ret == -1) {
-			return -errno;
-
-		} else if (ret == 0) {
-			int elapsed_ms;
-
-			ret = clock_gettime(CLOCK_MONOTONIC, &time_end);
-			if (ret == -1) {
-				fprintf(stderr, "Failed to read time: %s\n",
-						strerror(errno));
-				return -errno;
-			}
-
-			elapsed_ms = time_diff_ms(&time_start, &time_end);
-			if (elapsed_ms >= timeout_ms) {
-				fprintf(stderr, "Timed out waiting for read\n");
-				return -ETIMEDOUT;
-			}
-			timeout_ms -= elapsed_ms;
-		}
-
-		chunk_read = read(fd, (uint8_t *)data + total_read, size);
-		if (chunk_read == -1) {
-			return -errno;
-		}
-
-		total_read += chunk_read;
-	}
-
-	return total_read;
-}
-
-static int bl_cmd_read_message(
-		int dev_fd,
-		int timeout_ms,
-		union bl_msg_data *msg)
-{
-	ssize_t expected_len;
-	ssize_t read_len;
-
-	expected_len = sizeof(msg->type);
-	read_len = bl_read(dev_fd, &msg->type, expected_len, timeout_ms);
-	if (read_len != expected_len) {
-		fprintf(stderr, "Failed to read message type from device");
-		if (read_len < 0)
-			fprintf(stderr, ": %s", strerror(-read_len));
-		fprintf(stderr, "\n");
-		return -read_len;
-	}
-
-	expected_len = bl_msg_type_to_len(msg->type) - sizeof(msg->type);
-	read_len = bl_read(dev_fd, &msg->type + sizeof(msg->type),
-			expected_len, timeout_ms);
-	if (read_len != expected_len) {
-		fprintf(stderr, "Failed to read message body from device");
-		if (read_len < 0)
-			fprintf(stderr, ": %s", strerror(-read_len));
-		fprintf(stderr, "\n");
-		return -read_len;
-	}
-
-	if ((msg->type == BL_MSG_SAMPLE_DATA16) ||
-			(msg->type == BL_MSG_SAMPLE_DATA32)) {
-
-		size_t sample_size = msg->type == BL_MSG_SAMPLE_DATA32 ?
-				sizeof(uint32_t) : sizeof(uint16_t);
-
-		expected_len = sample_size * msg->sample_data.count;
-		read_len = bl_read(dev_fd, msg->sample_data.data16,
-				expected_len, timeout_ms);
-
-		if (read_len != expected_len) {
-			fprintf(stderr, "Failed to read %"PRIu8" samples from device",
-					msg->sample_data.count);
-			if (read_len < 0)
-				fprintf(stderr, ": %s", strerror(-read_len));
-			fprintf(stderr, "\n");
-			return -read_len;
-		}
-	}
-
-	return 0;
-}
-
 int bl_cmd_read_and_print_message(int dev_fd, int timeout_ms)
 {
-	union bl_msg_data msg = { 0 };
-	int ret = bl_cmd_read_message(dev_fd, timeout_ms, &msg);
-	if (ret == 0) {
+	union bl_msg_data msg;
+
+	if (bl_msg_read(dev_fd, timeout_ms, &msg)) {
 		bl_msg_yaml_print(stdout, &msg);
 		if (msg.type == BL_MSG_RESPONSE) {
 			if (msg.response.error_code != BL_ERROR_NONE) {
@@ -209,7 +91,7 @@ int bl_cmd_read_and_print_message(int dev_fd, int timeout_ms)
 		}
 	}
 
-	return ret;
+	return 0;
 }
 
 static int bl_open_device(const char *dev_path)
@@ -463,9 +345,7 @@ int bl_cmd_receive_and_print_loop(int dev_fd)
 	int ret = EXIT_SUCCESS;
 	do {
 		ret = bl_cmd_read_and_print_message(dev_fd, 10000);
-		if (ret == EINTR) {
-			bl_sig_killed = true;
-		} else if (ret == ECONNABORTED) {
+		if (ret == ECONNABORTED) {
 			return EXIT_SUCCESS;
 		} else if (ret != 0) {
 			ret = EXIT_FAILURE;
