+++ tools/msg.h
@@ -22,4 +22,17 @@ void bl_msg_yaml_print(
 		FILE *file,
 		const union bl_msg_data *msg);
 
+/**
+ * Read raw msg from given file descriptor into given message data structure.
+ *
+ * \param[in]     fd       File descriptor to read message from.
+ * \param[in]     timeout  Timeout in ms.
+ * \param[in,out] msg      Message structure to populate.
+ * \return true on success, false otherwise.
+ */
+bool bl_msg_read(
+		int fd,
+		int timeout,
+		union bl_msg_data *msg);
+
 #endif /* BL_TOOLS_MSG_H */
